+++ packages/kit/src/runtime/client/remote-functions/form.svelte.js
@@ -237,12 +237,7 @@ export function form(id) {
 
 				event.preventDefault();
 
-				const form_data = new FormData(form);
-
-				const submitter_name = event.submitter?.getAttribute('name');
-				if (submitter_name) {
-					form_data.append(submitter_name, event.submitter?.getAttribute('value') ?? '');
-				}
+				const form_data = new FormData(form, event.submitter);
 
 				if (DEV) {
 					validate_form_data(form_data, clone(form).enctype);
@@ -352,7 +347,7 @@ export function form(id) {
 				event.stopPropagation();
 				event.preventDefault();
 
-				const form_data = new FormData(form);
+				const form_data = new FormData(form, target);
 
 				if (DEV) {
 					const enctype = target.hasAttribute('formenctype')
@@ -360,10 +355,6 @@ export function form(id) {
 						: clone(form).enctype;
 
 					validate_form_data(form_data, enctype);
-				}
-
-				if (target.name) {
-					form_data.append(target.name, target?.getAttribute('value') ?? '');
 				}
 
 				await handle_submit(form, form_data, callback);
@@ -439,12 +430,7 @@ export function form(id) {
 
 					const id = ++validate_id;
 
-					const form_data = new FormData(element);
-
-					const submitter_name = submitter?.getAttribute('name');
-					if (submitter_name) {
-						form_data.append(submitter_name, submitter?.getAttribute('value') ?? '');
-					}
+					const form_data = new FormData(element, submitter);
 
 					/** @type {readonly StandardSchemaV1.Issue[]} */
 					let array = [];